+++ src/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 v.4.7.2
 * [bug] Small glitch in "hook" page in help admin
+* [bug] Unticking "Cache auto refresh" in cache admin was unticking "Cache expire" if set on "never" (removed js update in admin-cache-options template)
 * [technical] No automatized refresh anymore when visiting admin cache "manage cache" (edited Cache_Files_Management class)
 
 v.4.7.1